Tcc 9 forum

+++ backend/services/question.service.js
@@ -1,40 +1,43 @@
+/* eslint-disable no-unused-vars */
 const httpStatus = require('http-status');
 const { prisma } = require('../config/database');
 const ApiError = require('../utils/ApiError');
 
-const createQuestion = async (questionBody, authorId, tags) => {
+const createQuestion = async (questionBody, authorId, tagId) => {
   const question = await prisma.question.create({
     data: {
       ...questionBody,
       user: {
         connect: { id: authorId },
       },
+      tag: {
+        connect: { id: tagId },
+      },
     },
   });
-  if (Array.isArray(tags)) {
-    await Promise.all(
-      tags.map((tagId) =>
-        prisma.questionTags.create({
-          data: {
-            questionId: question.id,
-            tagId,
-          },
-        }),
-      ),
-    );
-  }
 
   return question;
 };
 
 const queryQuestions = async (filter, options) => {
   const questions = await prisma.question.findMany({
-    where: filter,
+    where: {
+      title: {
+        contains: filter.title,
+        mode: 'insensitive',
+      },
+      tag: { name: filter.tagName },
+      user: {
+        id: filter.userId,
+      },
+    },
     take: options.limit,
     skip: options.skip,
-    orderBy: options.sort,
+    orderBy: { createdAt: options.sortBy },
     include: {
       Answer: true,
+      tag: true,
+      user: true,
     },
   });
 
@@ -46,6 +49,8 @@ const getQuestionById = async (questionId) => {
     where: { id: questionId },
     include: {
       Answer: true,
+      tag: true,
+      user: true,
     },
   });
 
@@ -56,35 +61,18 @@ const getQuestionById = async (questionId) => {
   return question;
 };
 
-const updateQuestionById = async (userId, questionId, updateBody, newTags) => {
+const updateQuestionById = async (userId, questionId, updateBody) => {
   const question = await getQuestionById(questionId);
-  const user = await prisma.user.findUnique({
+  const user = await prisma.user.findFirst({
     where: { id: userId },
   });
 
-  if (!user) {
+  if (!user || question.userId !== userId) {
     throw new ApiError(httpStatus.FORBIDDEN, 'Você não tem permissão para acessar este recurso!');
   }
 
   if (question.locked && user.role !== 'ADMIN') {
     throw new ApiError(httpStatus.BAD_REQUEST, 'Esta pergunta já foi respondida e não pode ser modificada.');
-  }
-
-  await prisma.questionTags.deleteMany({
-    where: { questionId },
-  });
-
-  if (Array.isArray(newTags)) {
-    await Promise.all(
-      newTags.map((tagId) =>
-        prisma.questionTags.create({
-          data: {
-            questionId: question.id,
-            tagId,
-          },
-        }),
-      ),
-    );
   }
 
   return prisma.question.update({
@@ -93,12 +81,11 @@ const updateQuestionById = async (userId, questionId, updateBody, newTags) => {
   });
 };
 
-const deleteQuestionById = async (questionId) => {
-  await getQuestionById(questionId);
-
-  await prisma.questionTags.deleteMany({
-    where: { questionId },
-  });
+const deleteQuestionById = async (questionId, userId) => {
+  const question = await getQuestionById(questionId);
+  if (question.userId !== userId) {
+    throw new ApiError(httpStatus.FORBIDDEN, 'Você não tem permissão para acessar este recurso!');
+  }
 
   return prisma.question.delete({
     where: { id: questionId },
